style signup modal and fixed bugs on login modal
