+++ frontend/src/components/Navigation/Navigation.jsx
@@ -1,10 +1,7 @@
 import { NavLink } from "react-router-dom";
 import { useSelector } from "react-redux";
 import ProfileButton from "./ProfileButton";
-import OpenModalButton from "../OpenModalButton";
-import LoginFormModal from "../LoginFormModal";
 import "./Navigation.css";
-import SignupFormModal from "../SignupFormModal/SignupFormPage";
 
 const Navigation = ({ isLoaded }) => {
   const sessionUser = useSelector((state) => state.session.user);
