--- frontend/src/components/SignupFormModal/SignupFormPage.jsx
@@ -1,119 +1,0 @@
-import { useState } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import "./SignupFormModal.css";
-import { Navigate } from "react-router-dom";
-import * as sessionActions from "../../store/session";
-import { useModal } from "../../context/Modal";
-
-const SignupFormModal = () => {
-  const sessionUser = useSelector((state) => state.session.user);
-  const dispatch = useDispatch();
-
-  const [firstName, setFirstName] = useState("");
-  const [lastName, setLastName] = useState("");
-  const [email, setEmail] = useState("");
-  const [username, setUsername] = useState("");
-  const [password, setPassword] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-
-  const [errors, setErrors] = useState({});
-  const { closeModal } = useModal();
-
-  if (sessionUser) return <Navigate to="/" replace />;
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    if (password !== confirmPassword) {
-      setErrors({ confirmPassword: "Passwords do not match" });
-      return;
-    }
-
-    setErrors({});
-
-    const user = {
-      firstName,
-      lastName,
-      email,
-      username,
-      password
-    };
-
-    dispatch(sessionActions.registerUser(user)).then(closeModal).catch(async (res) => {
-      const data = await res.json();
-      if (data?.errors) setErrors(data.errors);
-    });
-  };
-
-  return (
-    <>
-      <h1>Sign Up To Sleep Inn!</h1>
-      <form className="signup-form" onSubmit={handleSubmit}>
-        <label>
-          First Name
-          <input
-            type="text"
-            value={firstName}
-            onChange={(e) => setFirstName(e.target.value)}
-            required
-          />
-        </label>
-        {errors.firstName && <p className="error">{errors.firstName}</p>}
-        <label>
-          Last Name
-          <input
-            type="text"
-            value={lastName}
-            onChange={(e) => setLastName(e.target.value)}
-            required
-          />
-        </label>
-        {errors.lastName && <p className="error">{errors.lastName}</p>}
-        <label>
-          Email
-          <input
-            type="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            required
-          />
-        </label>
-        {errors.email && <p className="error">{errors.email}</p>}
-        <label>
-          Username
-          <input
-            type="text"
-            value={username}
-            onChange={(e) => setUsername(e.target.value)}
-            required
-          />
-        </label>
-        {errors.username && <p className="error">{errors.username}</p>}
-        <label>
-          Password
-          <input
-            type="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            required
-          />
-        </label>
-        {errors.password && <p className="error">{errors.password}</p>}
-        <label>
-          Confirm Password
-          <input
-            type="password"
-            value={confirmPassword}
-            onChange={(e) => setConfirmPassword(e.target.value)}
-            required
-          />
-        </label>
-        {errors.confirmPassword && (
-          <p className="error">{errors.confirmPassword}</p>
-        )}
-        <button type="submit">Sign Up</button>
-      </form>
-    </>
-  );
-};
-
-export default SignupFormModal;
